Optimize code

+++ publish.go
@@ -1,23 +1,23 @@
 package microsite
 
 import (
-	"context"
+	"errors"
 	"fmt"
 	"reflect"
 	"strings"
 	"time"
 
 	"github.com/jinzhu/gorm"
+	"github.com/qor/admin"
 	"github.com/qor/media/oss"
 	"github.com/qor/publish2"
 	"github.com/theplant/gormutils"
 )
 
-func Publish(ctx context.Context, version QorMicroSiteInterface, printActivityLog bool) (err error) {
-	_db := ctx.Value("DB").(*gorm.DB)
-	tableName := _db.NewScope(version).TableName()
+func Publish(db *gorm.DB, version QorMicroSiteInterface, printActivityLog bool) (err error) {
+	tableName := db.NewScope(version).TableName()
 
-	err = gormutils.Transact(_db, func(tx *gorm.DB) (err1 error) {
+	err = gormutils.Transact(db, func(tx *gorm.DB) (err1 error) {
 		defer func() {
 			if err1 != nil {
 				eventType := fmt.Sprintf("%s:PublishingError", strings.Title(tableName))
@@ -25,11 +25,20 @@ func Publish(ctx context.Context, version QorMicroSiteInterface, printActivityLo
 			}
 		}()
 
+		// Find possible online version
 		iRecord := reflect.New(reflect.TypeOf(version).Elem()).Interface()
-		admDB.Set(publish2.VersionMode, publish2.VersionMultipleMode).Set(publish2.ScheduleMode, publish2.ModeOff).
+		if err1 = tx.Set(admin.DisableCompositePrimaryKeyMode, "on").Set(publish2.VersionMode, publish2.VersionMultipleMode).Set(publish2.ScheduleMode, publish2.ModeOff).
 			Where("id = ? AND status = ?", version.GetId(), Status_published).Where("version_name <> ?", version.GetVersionName()).
-			First(iRecord)
-		liveRecord := iRecord.(QorMicroSiteInterface)
+			First(iRecord).Error; err1 != nil && err1 != gorm.ErrRecordNotFound {
+			return
+		}
+
+		liveRecord, ok := iRecord.(QorMicroSiteInterface)
+		if !ok {
+			return errors.New("given record doesn't implement QorMicroSiteInterface")
+		}
+
+		// If there is a published version, unpublish it
 		if liveRecord.GetId() != 0 {
 			for _, o := range liveRecord.GetFilesPathWithSiteURL() {
 				oss.Storage.Delete(o)
@@ -40,14 +49,20 @@ func Publish(ctx context.Context, version QorMicroSiteInterface, printActivityLo
 				return
 			}
 
-			if err1 = liveRecord.UnPublishCallBack(_db, liveRecord.GetMicroSiteURL()); err1 != nil {
+			if err1 = liveRecord.UnPublishCallBack(tx, liveRecord.GetMicroSiteURL()); err1 != nil {
 				return
 			}
 		}
 
+		// Publish given version
 		version.SetStatus(Status_published)
 		version.SetVersionPriority(fmt.Sprintf("%v", time.Now().UTC().Format(time.RFC3339)))
 		if err1 = tx.Save(version).Error; err1 != nil {
+			return
+		}
+
+		// If callback has error, instead of rollback s3 changes. we call that expensive operation later.
+		if err1 = version.PublishCallBack(tx, version.GetMicroSiteURL()); err1 != nil {
 			return
 		}
 
@@ -55,7 +70,7 @@ func Publish(ctx context.Context, version QorMicroSiteInterface, printActivityLo
 			return
 		}
 
-		return version.PublishCallBack(_db, version.GetMicroSiteURL())
+		return
 	})
 
 	return
