+++ republish.go
@@ -1,23 +1,23 @@
 package microsite
 
 import (
-	"context"
+	"errors"
 	"fmt"
 	"reflect"
 	"strings"
 	"time"
 
 	"github.com/jinzhu/gorm"
+	"github.com/qor/admin"
 	"github.com/qor/media/oss"
 	"github.com/qor/publish2"
 	"github.com/theplant/gormutils"
 )
 
-func Republish(ctx context.Context, version QorMicroSiteInterface, printActivityLog bool) (err error) {
-	_db := ctx.Value("DB").(*gorm.DB)
-	tableName := _db.NewScope(version).TableName()
+func Republish(db *gorm.DB, version QorMicroSiteInterface, printActivityLog bool) (err error) {
+	tableName := db.NewScope(version).TableName()
 
-	err = gormutils.Transact(_db, func(tx *gorm.DB) (err1 error) {
+	err = gormutils.Transact(db, func(tx *gorm.DB) (err1 error) {
 		defer func() {
 			if err1 != nil {
 				eventType := fmt.Sprintf("%s:RepublishingError", strings.Title(tableName))
@@ -26,9 +26,16 @@ func Republish(ctx context.Context, version QorMicroSiteInterface, printActivity
 		}()
 
 		iRecord := reflect.New(reflect.TypeOf(version).Elem()).Interface()
-		admDB.Set(publish2.VersionMode, publish2.VersionMultipleMode).Set(publish2.ScheduleMode, publish2.ModeOff).
-			Where("id = ? AND status = ?", version.GetId(), Status_published).First(iRecord)
-		liveRecord := iRecord.(QorMicroSiteInterface)
+		if err1 = db.Set(admin.DisableCompositePrimaryKeyMode, "on").Set(publish2.VersionMode, publish2.VersionMultipleMode).Set(publish2.ScheduleMode, publish2.ModeOff).
+			Where("id = ? AND status = ?", version.GetId(), Status_published).First(iRecord).Error; err1 != nil && err1 != gorm.ErrRecordNotFound {
+			return
+		}
+
+		liveRecord, ok := iRecord.(QorMicroSiteInterface)
+		if !ok {
+			return errors.New("given record doesn't implement QorMicroSiteInterface")
+		}
+
 		if liveRecord.GetId() != 0 {
 			for _, o := range liveRecord.GetFilesPathWithSiteURL() {
 				oss.Storage.Delete(o)
@@ -40,7 +47,7 @@ func Republish(ctx context.Context, version QorMicroSiteInterface, printActivity
 			}
 
 			if liveRecord.GetVersionName() != version.GetVersionName() {
-				if err1 = liveRecord.UnPublishCallBack(_db, liveRecord.GetMicroSiteURL()); err1 != nil {
+				if err1 = liveRecord.UnPublishCallBack(db, liveRecord.GetMicroSiteURL()); err1 != nil {
 					return
 				}
 			}
@@ -57,7 +64,7 @@ func Republish(ctx context.Context, version QorMicroSiteInterface, printActivity
 		}
 
 		if liveRecord.GetVersionName() != version.GetVersionName() {
-			err1 = version.PublishCallBack(_db, version.GetMicroSiteURL())
+			err1 = version.PublishCallBack(db, version.GetMicroSiteURL())
 		}
 		return
 	})
