+++ unpublish.go
@@ -1,7 +1,6 @@
 package microsite
 
 import (
-	"context"
 	"fmt"
 	"strings"
 
@@ -10,11 +9,10 @@ import (
 	"github.com/theplant/gormutils"
 )
 
-func Unpublish(ctx context.Context, version QorMicroSiteInterface, printActivityLog bool) (err error) {
-	_db := ctx.Value("DB").(*gorm.DB)
-	tableName := _db.NewScope(version).TableName()
+func Unpublish(db *gorm.DB, version QorMicroSiteInterface, printActivityLog bool) (err error) {
+	tableName := db.NewScope(version).TableName()
 
-	err = gormutils.Transact(_db, func(tx *gorm.DB) (err1 error) {
+	err = gormutils.Transact(db, func(tx *gorm.DB) (err1 error) {
 		defer func() {
 			if err1 != nil {
 				eventType := fmt.Sprintf("%s:UnpublishingError", strings.Title(tableName))
@@ -27,12 +25,15 @@ func Unpublish(ctx context.Context, version QorMicroSiteInterface, printActivity
 			return
 		}
 
+		if err1 = version.UnPublishCallBack(tx, version.GetMicroSiteURL()); err1 != nil {
+			return
+		}
+
 		for _, o := range version.GetFilesPathWithSiteURL() {
 			oss.Storage.Delete(o)
 		}
 
-		return version.UnPublishCallBack(_db, version.GetMicroSiteURL())
-
+		return
 	})
 
 	return
